msm: kgsl: Prevent missing servicing of events waiting on timestamps
Prevent instances where we miss servicing an event which was waiting
on a timestamp. We do this by looping in the events list, until
we finish servicing all the members of the list, and if we come across
an event which has not been serviced yet, we write a dummy interrupt to
the ringbuffer so that we don't miss when that timestamp's work is done.

Change-Id: Ib7b3554efa0d9a8bdc20d1722d912977f72f3ed5
Signed-off-by: Harsh Vardhan Dwivedi <hdwivedi@codeaurora.org>